+++ src/main/java/org/realityforge/replicant/client/gwt/WebPollerDataLoaderService.java
@@ -39,6 +39,11 @@ public abstract class WebPollerDataLoaderService<T extends ClientSession<T, G>, 
       final RequestBuilder rb = new RequestBuilder( RequestBuilder.GET, getPollURL() );
       rb.setHeader( ReplicantContext.SESSION_ID_HEADER, getSessionID() );
       rb.setHeader( "Pragma", "no-cache" );
+      final String authenticationToken = getSessionContext().getAuthenticationToken();
+      if ( null != authenticationToken )
+      {
+        rb.setHeader( "Authorization", "Bearer " + authenticationToken );
+      }
       return rb;
     }
   }